moving centroid configs to separate file
