+++ projects/dendrites/permutedMNIST/experiments/base.py
@@ -34,23 +34,13 @@ from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
 )
-from nupic.research.frameworks.pytorch.datasets import (
-    ContextDependentPermutedMNIST,
-    PermutedMNIST,
-)
-from nupic.research.frameworks.vernon import ContinualLearningExperiment, mixins
+from nupic.research.frameworks.pytorch.datasets import ContextDependentPermutedMNIST
+from nupic.research.frameworks.vernon import mixins
 
 
 class PermutedMNISTExperiment(mixins.RezeroWeights,
                               mixins.PermutedMNISTTaskIndices,
                               DendriteContinualLearningExperiment):
-    pass
-
-
-class CentroidExperiment(mixins.RezeroWeights,
-                         mixins.CentroidContext,
-                         mixins.PermutedMNISTTaskIndices,
-                         ContinualLearningExperiment):
     pass
 
 
@@ -207,61 +197,6 @@ BASE_50_SPARSITY_SEARCH.update(
     ),
 )
 
-# Centroid method for inferring contexts: 10 permutedMNIST tasks
-CENTROID_10 = dict(
-    experiment_class=CentroidExperiment,
-    num_samples=8,
-
-    # Results path
-    local_dir=os.path.expanduser("~/nta/results/experiments/dendrites"),
-
-    dataset_class=PermutedMNIST,
-    dataset_args=dict(
-        num_tasks=10,
-        root=os.path.expanduser("~/nta/results/data/"),
-        download=False,  # Change to True if running for the first time
-        seed=42,
-    ),
-
-    model_class=DendriticMLP,  # CentroidDendriticMLP does not affect accuracy..??
-    model_args=dict(
-        input_size=784,
-        output_size=10,  # Single output head shared by all tasks
-        hidden_sizes=[2048, 2048],
-        num_segments=10,
-        dim_context=784,
-        kw=True,
-        kw_percent_on=0.1,
-        dendrite_weight_sparsity=0.0,
-        weight_sparsity=0.5,
-        context_percent_on=0.1,
-    ),
-
-    batch_size=256,
-    val_batch_size=512,
-    epochs=3,
-    tasks_to_validate=(0, 1, 4, 9, 24, 49),
-    num_tasks=10,
-    num_classes=10 * 10,
-    distributed=False,
-    seed=tune.sample_from(lambda spec: np.random.randint(2, 10000)),
-
-    loss_function=F.cross_entropy,
-    optimizer_class=torch.optim.Adam,  # On permutedMNIST, Adam works better than
-                                       # SGD with default hyperparameter settings
-    optimizer_args=dict(lr=5e-4),
-)
-
-# Centroid method for inferring contexts: 50 permutedMNIST tasks
-CENTROID_50 = deepcopy(CENTROID_10)
-CENTROID_50["dataset_args"].update(num_tasks=50)
-CENTROID_50["model_args"].update(num_segments=50)
-CENTROID_50.update(
-    epochs=2,
-    num_tasks=50,
-    num_classes=10 * 50,
-)
-
 
 # Export configurations in this file
 CONFIGS = dict(
@@ -270,6 +205,4 @@ CONFIGS = dict(
     base_10_sparsity=BASE_10_SPARSITY_SEARCH,
     base_50_search=BASE_50_SPARSITY_SEARCH,
     base2=BASE2,
-    centroid_10=CENTROID_10,
-    centroid_50=CENTROID_50
 )
